+++ lib/shared/input_models/input_models.dart
@@ -1,2 +1,1 @@
 export 'email.dart';
-export 'password.dart';
