+++ lib/sign_in/input_models/password.dart
@@ -1,0 +1,23 @@
+import 'package:formz/formz.dart';
+
+/// Validation errors for the [Password] [FormzInput].
+enum PasswordValidationError {
+  /// Field is empty error.
+  empty
+}
+
+/// {@template password}
+/// Form input for a sing-in password input.
+/// {@endtemplate}
+class Password extends FormzInput<String, PasswordValidationError> {
+  /// {@macro password}
+  const Password.pure() : super.pure('');
+
+  /// {@macro password}
+  const Password.dirty([super.value = '']) : super.dirty();
+
+  @override
+  PasswordValidationError? validator(String? value) {
+    return value?.isEmpty ?? true ? PasswordValidationError.empty : null;
+  }
+}
