+++ lib/sign_up/input_models/input_models.dart
@@ -2,4 +2,5 @@ export 'birthdate.dart';
 export 'confirmed_password.dart';
 export 'genre.dart';
 export 'name.dart';
+export 'password.dart';
 export 'username.dart';
